Start polishing
