+++ src/dcegm/pre_processing.py
@@ -6,8 +6,6 @@ from typing import Tuple
 import numpy as np
 import pandas as pd
 from dcegm.aggregate_policy_value import calc_current_period_value
-from dcegm.aggregate_policy_value import calc_expected_value
-from dcegm.aggregate_policy_value import calc_next_period_choice_probs
 
 
 def get_partial_functions(
@@ -40,13 +38,6 @@ def get_partial_functions(
         beta=params.loc[("beta", "beta"), "value"],
         compute_utility=compute_utility,
     )
-    compute_expected_value = partial(
-        calc_expected_value,
-        params=params,
-    )
-    compute_next_choice_probs = partial(
-        calc_next_period_choice_probs, params=params, options=options
-    )
     compute_next_wealth_matrices = partial(
         user_budget_constraint,
         savings_grid=exogenous_savings_grid,
@@ -70,8 +61,6 @@ def get_partial_functions(
         compute_marginal_utility,
         compute_inverse_marginal_utility,
         compute_current_value,
-        compute_expected_value,
-        compute_next_choice_probs,
         compute_next_wealth_matrices,
         compute_next_marginal_wealth,
         store_current_policy_and_value,
